server: comment out more unused code, add mutex for subscriptions, FIX hanging bug after 30 min after hitting channel limit
the subscriber_worker() was adding subscription claims to the claims channel but the claims worker was commented out and not taking things off the channel.  The channel buffer size was 100 and it took about 30 min of regular use before the buffer limit was hit, causing all new subscriptions to hang indefintely. Lesson learned: be aware that adding to a channel might block if consumer of the channel isn't running or is too slow.